fix(guild): small file renames and fixes
