+++ capy_discord/exts/guild/_schemas.py
